% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm.get.summary.R
\name{palm.get.summary}
\alias{palm.get.summary}
\title{Generate PALM summary statistics for individual studies}
\usage{
palm.get.summary(
  null.obj,
  covariate.interest,
  cluster = NULL,
  correct = "median"
)
}
\arguments{
\item{null.obj}{The output object from the \code{palm.null.model} function.}

\item{...}{Additional arguments passed to \code{palm}.
In \code{covariate.interest} and \code{cluster}, the order of samples must match
the order in \code{rel.abd} used in the \code{palm.null.model} function.}
}
\value{
A list containing one component per study. Each component includes the following elements:
\item{est}{A matrix containing the association effect estimates for the study.
Rows correspond to microbial feature IDs and columns correspond to covariate IDs.}
\item{stderr}{A matrix containing the standard errors of the association effect estimates for the study.
Rows correspond to microbial feature IDs and columns correspond to covariate IDs.}
\item{n}{The sample size for the study.}
}
\description{
This function takes the output object from the \code{palm.null.model} function and
constructs summary statistics for the covariates of interest in each study.
The resulting output can be directly used as input to the \code{palm.meta.summary} function
to perform meta-analysis.
}
\examples{
\donttest{
library(PALM)
data("CRC_data", package = "PALM")
CRC_abd <- CRC_data$CRC_abd
CRC_meta <- CRC_data$CRC_meta

########## Generate summary statistics ##########
rel.abd <- list()
covariate.interest <- list()
for (d in unique(CRC_meta$Study)) {
  rel.abd[[d]] <- CRC_abd[CRC_meta$Sample_ID[CRC_meta$Study == d], ]
  disease <- as.numeric(CRC_meta$Group[CRC_meta$Study == d] == "CRC")
  names(disease) <- CRC_meta$Sample_ID[CRC_meta$Study == d]
  covariate.interest[[d]] <- matrix(disease, ncol = 1,
                                    dimnames = list(names(disease), "disease"))
}

null.obj <- palm.null.model(rel.abd = rel.abd)
summary.stats <- palm.get.summary(null.obj = null.obj,
                                  covariate.interest = covariate.interest)
}
}
\seealso{
\code{\link{palm.null.model}},
\code{\link{palm.meta.summary}},
\code{\link{palm}}
}
