% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm.null.model.R
\name{palm.null.model}
\alias{palm.null.model}
\title{Get information from the null model}
\usage{
palm.null.model(
  rel.abd,
  covariate.adjust = NULL,
  depth = NULL,
  depth.filter = 0,
  prev.filter = 0.1
)
}
\arguments{
\item{...}{Additional arguments passed to \code{palm}.}
}
\value{
A list containing one component per study. Each component includes the following elements:
\item{Y_I}{A matrix of predicted abundances.}
\item{Y_R}{A matrix of abundance residuals.}
\item{Z}{The cleaned and formatted \code{covariate.adjust}.}
\item{rm.sample.idx}{Indices of the samples removed from the study.}
}
\description{
The \code{palm.null.model} function computes the estimated mean microbial feature
proportions and residuals of the relative abundances under the null model for each study.
This function does not require the \code{covariate.interest} information.
The output of this function is used as input for the \code{palm.get.summary} function
to construct summary statistics for the covariates of interest in each study.
}
\examples{
\donttest{
library(PALM)
data("CRC_data", package = "PALM")
CRC_abd <- CRC_data$CRC_abd
CRC_meta <- CRC_data$CRC_meta

########## Generate summary statistics ##########
rel.abd <- list()
for (d in unique(CRC_meta$Study)) {
  rel.abd[[d]] <- CRC_abd[CRC_meta$Sample_ID[CRC_meta$Study == d], ]
}

null.obj <- palm.null.model(rel.abd = rel.abd)
}
}
\seealso{
\code{\link{palm.get.summary}},
\code{\link{palm.meta.summary}},
\code{\link{palm}}
}
