% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm.get.summary.R
\name{palm.get.summary}
\alias{palm.get.summary}
\title{Generate palm summary statistics for individual studies.}
\usage{
palm.get.summary(null.obj, covariate.interest, cluster = NULL)
}
\arguments{
\item{null.obj}{The output of function "palm.null.model".}

\item{...}{See function palm in "covariate.interest" and "cluster",
the order of samples should be matched with the order in "rel.abd" used in the "palm.null.model" function.}
}
\value{
Output a list with each component for a study. The component includes the following elements.
\item{est}{A matrix contains the relative-abundance association coefficient estimates for the study. The row names are microbial feature IDs
and the column names are the covariates of interest IDs.}
\item{var}{A matrix contains the variances of the relative-abundance association coefficient estimates for the study.
The row names are microbial feature IDs
and the column names are the covariates of interest IDs.}
\item{n}{Sample size for the study.}
}
\description{
This function directly takes the output object from the "palm.null.model" function and
constructs summary statistics for covariates of interest in each study.
The output can be directly used by the function "plam.meta.summary" to perform meta-analysis.
}
\examples{
\donttest{
library("PALM")
data("CRC_data", package = "PALM")
CRC_abd <- CRC_data$CRC_abd
CRC_meta <- CRC_data$CRC_meta

########## Generate summary statistics ##########
rel.abd <- list()
covariate.interest <- list()
for(d in unique(CRC_meta$Study)){
  rel.abd[[d]] <- CRC_abd[CRC_meta$Sample_ID[CRC_meta$Study == d],]
  disease <- as.numeric(CRC_meta$Group[CRC_meta$Study == d] == "CRC")
  names(disease) <- CRC_meta$Sample_ID[CRC_meta$Study == d]
  covariate.interest[[d]] <- matrix(disease, ncol = 1, dimnames = list(names(disease), "disease"))
}

 null.obj <- palm.null.model(rel.abd = rel.abd)

 summary.stats <- palm.get.summary(null.obj = null.obj, covariate.interest = covariate.interest)
}

}
\seealso{
\code{\link{palm.null.model}},
\code{\link{palm.meta.summary}},
\code{\link{palm}}
}
