% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm.null.model.R
\name{palm.null.model}
\alias{palm.null.model}
\title{Get information from the null model.}
\usage{
palm.null.model(
  rel.abd,
  covariate.adjust = NULL,
  N = NULL,
  depth.filter = 0,
  prev.filter = 0.1
)
}
\arguments{
\item{...}{See function palm}
}
\value{
Output a list with each component for a study. The component includes the following elements.
\item{Y_I}{A matrix of intermediate results.}
\item{Y_R}{A matrix of intermediate results.}
\item{Z}{A vector of microbiome sequencing depth for the study.}
\item{rm.sample.idx}{The index of the removed samples for the study.}
}
\description{
The "palm.null.model" function computes the estimated mean microbial feature proportions and residuals
of the relative abundance under the null model of no associations for each study (so this function does not need the
covariate.interest information). The output of this function will be fed into the "palm.get.summary" function to
construct summary statistics for covariates of interest in each study.
}
\examples{
\donttest{
library("PALM")
data("CRC_data", package = "miMeta")
CRC_abd <- CRC_data$CRC_abd
CRC_meta <- CRC_data$CRC_meta

########## Generate summary statistics ##########
rel.abd <- list()
for(d in unique(CRC_meta$Study)){
  rel.abd[[d]] <- CRC_abd[CRC_meta$Sample_ID[CRC_meta$Study == d],]
}

null.obj <- palm.null.model(rel.abd = rel.abd)
}

}
\seealso{
\code{\link{palm.get.summary}},
\code{\link{palm.test}},
\code{\link{palm}}
}
